feat: Add FavoriteLyricCard and FavoritesFilterBottomSheet widgets for enhanced favorites functionality
